Improve sighting location map
