+++ src/components/cards/GpsCard.jsx
@@ -1,14 +1,6 @@
 import React from 'react';
-import GoogleMapReact from 'google-map-react';
-
-import { googleMapsApiKey } from '../../constants/apiKeys';
+import SinglePoint from '../maps/SinglePoint';
 import Card from './Card';
-
-function createMapOptions() {
-  return {
-    gestureHandling: 'cooperative',
-  };
-}
 
 export default function GpsCard({
   title,
@@ -19,23 +11,7 @@ export default function GpsCard({
   return (
     <Card title={title} titleId={titleId}>
       <div style={{ height: 280, marginTop: 4 }}>
-        <GoogleMapReact
-          options={createMapOptions}
-          bootstrapURLKeys={{ key: googleMapsApiKey }}
-          defaultCenter={{
-            lat,
-            lng,
-          }}
-          defaultZoom={7}
-          yesIWantToUseGoogleMapApiInternals
-          onGoogleApiLoaded={({ map, maps }) => {
-            const markerPosition = new maps.LatLng(lat, lng);
-            const marker = new maps.Marker({
-              position: markerPosition,
-            });
-            marker.setMap(map);
-          }}
-        />
+        <SinglePoint lat={lat} lng={lng} />
       </div>
     </Card>
   );
